// src/ui/blockart/blockColors.generated.ts
export const TONE_BODY: Partial<Record<string, string>> = {
  acoustasonic: "#b49c84",
  blonde: "#ccb49c",
  brownface: "#846c54",
  jubilee: "#9c9c9c",
  orange: "#d8600c",
  tweed: "#e49c3c",
  wine: "#ccb49c",
};

export const CLOTH_BASE: Partial<Record<string, string>> = {
  acoustmesh: "#242424",
  basket: "#a88454",
  bassgrille: "#848484",
  black: "#0c0c0c",
  blackjute: "#242424",
  diamond: "#6c3c3c",
  evhgrille: "#0c0c0c",
  fliptopgrille: "#54543c",
  grey: "#48483c",
  jcgrey: "#3c3c3c",
  oxblood: "#242424",
  saltpep: "#181818",
  silver: "#848484",
  silverturq: "#848484",
  svtcloth: "#848484",
  tweed: "#543c24",
  wheat: "#b49c6c",
};